Minor refactor of tabs into their own pages.
